replace docker exec to host exec

+++ script/node-init/horus-agent-1.3.7/scripts/upsql_file.sh
@@ -20,7 +20,13 @@ if [ "${running_status}" != "true" ]; then
 	exit 4
 fi
 
-docker exec $INSTANCE mysql -S /DBAASDAT/upsql.sock mysql -u$USER -p$PASSWD  -e"show variables where Variable_name in ( 'log_error', 'slow_query_log_file');" >$VARFILE  2>/dev/null
+EXEC_BIN=`which mysql 2>/dev/null`
+if [ "${EXEC_BIN}" == '' ]; then
+	echo "not find mysql"
+	exit 4
+fi
+
+${EXEC_BIN} -S /${INSTANCE}_DAT_LV/upsql.sock mysql -u$USER -p$PASSWD  -e"show variables where Variable_name in ( 'log_error', 'slow_query_log_file');" >$VARFILE  2>/dev/null
 if [ $? -ne 0 ];then
 	echo "get variabes err"
 	rm  $STATUSFILE
